requirements.txt
# Core Gradio
gradio

# LLM Agent
ollama

# Video & Audio Processing
ffmpeg-python
soundfile

# Hugging Face AI/ML Libraries
torch
diffusers
transformers
accelerate
safetensors
huggingface_hub
datasets
scipy
Pillow